Improve `nil` placeholder rendering
